Upgrade node from 0.10 to 6 with many other package upgrades
First of all, I am sorry, there is not good reason to be here.

This commit embed a lot of package upgrades because it was hard to work more incrementally on that node upgrade.

Great thanks, kodus and all the rest to @Flightan who helped a lot with that migration.